Refactor : generate LadderGenerator and remove LineGenerator
